pyproject.toml
[project]
name = "mcp_massive"
version = "0.7.0"
description = "A MCP server project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.15.0",
    "httpx>=0.28.0",
    "bm25s>=0.2.0",
    "python-dotenv>=1.2.0",
]
[[project.authors]]
name = "Massive"
email = "support@massive.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.13.2",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "rust-just>=1.42.4",
]

[project.scripts]
mcp_massive = "mcp_massive:main"
